requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydub==0.25.1
soundfile==0.12.1
transformers==4.35.2
torch>=2.0.0
sentencepiece>=0.1.99
google-generativeai>=0.3.0
langchain>=0.0.300
langdetect>=1.0.9
language_tool_python>=2.7.0
pyttsx3>=2.90
openai-whisper>=20231117
faster-whisper>=0.10.0
streamlit==1.28.1
typing_extensions==4.8.0
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
prometheus-client==0.19.0
psutil==5.9.6
tenacity==8.2.3
requests==2.31.0
